examples/agents/python/aws-strands/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-genai==1.52.0
strands-agents==1.18.0
jinja2==3.1.6
python-dotenv==1.0.1
PyJWT==2.10.1
